Move some static pages to typescript
