+++ components/pages/staticPage/StaticPage.tsx
